pyproject.toml
[tool.poetry]
name = "sign-language-classifier"
version = "0.1.0"
description = "Sign language gesture classification with EfficientNet and PyTorch Lightning"
authors = ["Твоё Имя <you@example.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
pytorch-lightning = "^2.2.1"
torch = "^2.1.2"
torchvision = "^0.16.2"
timm = "^0.9.12"
hydra-core = "^1.3.2"
mlflow = "^2.10.1"
dvc = "^3.50.1"
dvc-gdrive = "^2.20.0"  # если используешь Google Drive
omegaconf = "^2.3.0"
numpy = "^1.26"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = "^24.4.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
